Create helpers.go
Perhaps a bit controversial, but in my application of the GCT, I use  it as a foundation library for an arbitrage app. I create "chains" of exchanges to computes spreads. I've had to deal with different symbols for BTC (e.g. Kraken uses XBT), and for it me it makes more sense to use the currency pairs in which the currency to sell is first currency, and buy currency second (so that a chain would look like for example EUR:BTC@Kraken->BTC:USD:Bitmex, instead of EUR:BTC@Kraken->USD:BTC@Bitmex.

The file also includes several functions to convert CurrencyPair & CurrencyItem to string